Moved cypress graphql support utils into better place.
